requirements.txt
tqdm
torch
torchvision
torchaudio
PyYAML
tensorboard
matplotlib
fire